Residual UNet
I tried to create a Residual UNet following: https://github.com/Kaido0/Brain-Tissue-Segment-Keras
